# docs/contracts/dre-narrative.openapi.yml
openapi: 3.0.3
info:
  title: DRE & Narrative — Aicfo API
  description: Contract for the dre-narrative module. Extracted from the canonical Aicfo OpenAPI spec.
  version: 0.2.0
servers:
  - url: http://localhost:3000
paths:
  /analysis/{analysisId}/dre:
    get:
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: analysisId
          required: true
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
        '400':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '401':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '403':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '404':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '409':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '422':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '500':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
  /analysis/{analysisId}/narrative:
    get:
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: analysisId
          required: true
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    cardType:
                      type: string
                    title:
                      type: string
                    body:
                      type: string
                    evidence:
                      type: array
                      items:
                        type: object
                        properties:
                          metric:
                            type: string
                          value:
                            type: number
                          unit:
                            type: string
                        required:
                          - metric
                          - value
                          - unit
                        additionalProperties: false
                    clientApproved:
                      type: boolean
                      nullable: true
                    clientComment:
                      type: string
                      nullable: true
                  required:
                    - id
                    - cardType
                    - title
                    - body
                    - evidence
                    - clientApproved
                    - clientComment
                  additionalProperties: false
        '400':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '401':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '403':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '404':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '409':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '422':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '500':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
  /analysis/{analysisId}/narrative/{cardId}/feedback:
    patch:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                approved:
                  type: boolean
                comment:
                  type: string
                  maxLength: 500
              required:
                - approved
              additionalProperties: false
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: analysisId
          required: true
        - schema:
            type: string
            format: uuid
          in: path
          name: cardId
          required: true
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                required:
                  - id
                additionalProperties: false
        '400':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '401':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '403':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '404':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '409':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '422':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
        '500':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                  instance:
                    type: string
                  requestId:
                    type: string
                required:
                  - type
                  - title
                  - status
                additionalProperties: false
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT